+++ chronicle-queue/src/main/java/net/openhft/chronicle/queue/impl/single/work/in/progress/SingleTailer.java
@@ -192,7 +192,7 @@ public class SingleTailer implements ExcerptTailer {
     }
 
     @Override
-    public boolean index(int cycle, long index) throws IOException {
+    public boolean index(long cycle, long index) throws IOException {
         return false;
     }
